# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Movie API",
    "description": "A simple API for managing movie data, integrated with a MySQL database using Knex.js. This API allows listing, searching, adding, updating, and deleting movie records, and includes basic user authentication.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "Retrieves a list of all registered users. (Note: In a production app, this would typically be an admin-only route.)",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "examples": {
                  "success": {
                    "summary": "Example User List",
                    "value": [
                      { "id": 1, "email": "user1@example.com" },
                      { "id": 2, "email": "user2@example.com" }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/profile": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user profile",
        "description": "Retrieves the profile details of the authenticated user. Requires a valid JWT token in the cookie.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Welcome, user@example.com" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No token or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Unauthorized: No token provided" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "User not found" }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "description": "User profile details",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "email": { "type": "string", "format": "email", "example": "user@example.com" }
        },
        "required": ["id", "email"]
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in an HttpOnly cookie"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "Error": { "type": "boolean", "example": true },
                "Message": { "type": "string", "example": "Error in MYSQL query or server failure" }
              }
            }
          }
        }
      }
    }
  }
}